Ryan

+++ src/components/Navigation.jsx
@@ -1,21 +1,31 @@
 import React, { Component } from "react";
 import {
   Navbar,
+  Row,
   Nav,
   Container,
   Form,
   FormControl,
   Button,
   Modal,
-  Image
+  Col,
+  Image,
 } from "react-bootstrap";
 import { Link, withRouter } from "react-router-dom";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Formik, Field, ErrorMessage } from "formik";
 import * as Yup from "yup";
 import axios from "axios";
 import Swal from "sweetalert2";
-import logo from "../assets/img/txtsplash.png";
+import logo from "../assets/img/moviebox.png";
+import { css } from "@emotion/core";
+import { RotateLoader } from "react-spinners/ClipLoader";
+import qs from "qs";
+
+const override = css`
+  display: block;
+  margin: 0 auto;
+  border-color: red;
+`;
 
 const schema = Yup.object().shape({
   email: Yup.string()
@@ -28,7 +38,7 @@ const schema = Yup.object().shape({
     .oneOf([Yup.ref("password"), null], "Passwords must match")
     .required("You forgot to type this field"),
   fullname: Yup.string()
-    .min(8, "Your name should be 8 characters long")
+    // .min(8, "Your name should be 8 characters long")
     .required("Name is required"),
   username: Yup.string()
     // .min(6, "Your username should be 6 characters long")
@@ -51,88 +61,98 @@ class Navigation extends Component {
     searchInput: "",
     redirect: false,
     token: localStorage.getItem("login"),
+    userLogin: localStorage.getItem("username"),
     loading: false,
+    // image: {
+    //   file: {},
+    //   url: "",
+    // },
   };
 
-  componentDidMount = () => {
-    const { token } = this.state;
+  // componentDidMount = () => {
+  //   const { token } = this.state;
 
-    if (token) {
-      this.getCurrentUser();
-    }
-  };
+  //   if (token) {
+  //     this.getCurrentUser();
+  //   }
+  // };
 
   componentDidUpdate = (prevProps, prevState) => {
-    const { token } = this.state;
+    const { userLogin } = this.state;
 
-    if (token !== prevState.token) {
-      if (token) {
-        this.getCurrentUser();
+    if (userLogin !== prevState.userLogin) {
+      if (userLogin) {
+        this.render();
       }
     }
   };
 
-  getCurrentUser = async () => {
-    try {
-      const { token } = this.state;
-      // console.log("token", token);
+  // getCurrentUser = async () => {
+  //   try {
+  //     const { token } = this.state;
+  //     const fetch = await axios.get("http://appdoto.herokuapp.com/api/user", {
+  //       headers: {
+  //         Authorization: token,
+  //       },
+  //     });
 
-      const fetch = await axios.get("http://appdoto.herokuapp.com/api/user", {
-        headers: {
-          Authorization: token,
-        },
-      });
-
-      console.log("fetch", fetch);
-      this.setState({
-        dataLoggedIn: fetch.data.data,
-      });
-    } catch (error) {
-      console.log("error", error);
-    }
-  };
+  //     console.log("fetch", fetch);
+  //     this.setState({
+  //       dataLoggedIn: fetch.data.data,
+  //     });
+  //   } catch (error) {
+  //     console.log("error", error);
+  //   }
+  // };
 
   onChange = (name, e) => {
     this.setState({
       [name]: e,
     });
-    // console.log(this.state);
   };
 
   login = async (values) => {
     try {
       const { email, password } = values;
-      // console.log(emailLogin, passLogin);
+      const dataLogin = qs.stringify({
+        email: email,
+        password: password
+      })
+      console.log(dataLogin);
       this.setState({ loading: true });
-      const submit = await axios.post(
-        "http://appdoto.herokuapp.com/api/users/login",
-        {
-          email,
-          password,
-        }
+      const submit = await axios({
+        method: "post",
+        url: "https://nameless-temple-74030.herokuapp.com/login",
+        data: dataLogin,
+        headers: {
+          "content-Type": "application/x-www-form-urlencoded",
+        },
+      }
       );
-      localStorage.setItem("login", submit.data.data.token);
-      this.onChange("token", submit.data.data.token);
+      localStorage.setItem("login", submit.data.access_token);
+      localStorage.setItem("idUser", submit.data.User.id);
+      localStorage.setItem("username", submit.data.User.username);
+      this.onChange("token", submit.data.access_token);
       this.setState({
         loading: false,
       });
       this.handleCloseLogin();
-      const username = submit.data.data.username;
+      const username = submit.data.User.username;
       Swal.fire({
         title: "Login Success",
         text: `Welcome ${username}`,
         icon: "success",
       });
     } catch (error) {
-      console.log("error ini", error);
+      console.log("error ini", error.response);
       this.setState({
         loading: false,
       });
-      const msg = error.response.data.messages.errors;
-      console.log(msg);
+      const msg = error.response.data.msg;
+      // console.log(msg);
       Swal.fire({
         title: "Login Failed",
-        text: "Email or password invalid!",
+        text: msg,
         icon: "error",
       });
     }
@@ -148,23 +168,39 @@ class Navigation extends Component {
     this.launchModalSign();
   };
 
-  signUp = async (values) => {
-    console.log("signupJalan", values);
+  signUp = async (values, images) => {
+    const { email, password, username, fullname } = values; //ini values dari formik
+    // console.log(formData);
+    console.log("value input", email, password, username, fullname);
+    const stringQs = qs.stringify({
+      email: email,
+      username: username,
+      password: password,
+      fullname: fullname,
+    });
+
+    console.log(stringQs);
     try {
-      const { email, password, username, fullname } = values; //ini values dari formik
-      const submit = await axios.post(
-        "http://appdoto.herokuapp.com/api/users/",
-        {
-          email: email,
-          password: password,
-          username: username,
-          fullname: fullname,
-          bio: "itulah bionya",
+      this.setState({ loading: true });
+      const submit = await axios({
+        method: "post",
+        url: "https://nameless-temple-74030.herokuapp.com/register",
+        data: stringQs,
+        headers: {
+          "content-Type": "application/x-www-form-urlencoded",
+        },
+      });
+      console.log(submit.data);
+      localStorage.setItem("login", submit.data.acces_token);
+      localStorage.setItem("idUser", submit.data.User.id);
+      localStorage.setItem("username", submit.data.User.username);
+      this.onChange("token", submit.data.acces_token);
+      this.setState({
+        loading: false,
+        dataLoggedIn: {
+          username : submit.data.User.username,
         }
-      );
-      localStorage.setItem("login", submit.data.data.token);
-      this.onChange("token", submit.data.data.token);
-      console.log(submit);
+      });
 
       this.handleCloseSign();
       Swal.fire({
@@ -174,26 +210,45 @@ class Navigation extends Component {
       });
     } catch (error) {
       console.log("error", error.response);
-      let { username, email } = error.response.data.errors;
-      console.log(username, email);
-      username = username ? `username ${username}` : "";
-      email = email ? `email ${email}` : "";
-
-      Swal.fire({
-        title: "Something went Wrong",
-
-        text: email + username,
-        icon: "error",
+      this.setState({
+        loading: false,
       });
+
+      const msg = error.response.data.msg;
+      msg ? (
+        Swal.fire({
+          title: "Something Went Wrong",
+          text: msg,
+          icon: "error",
+        })
+      ) : (
+        Swal.fire({
+          title: "Something Went Wrong",
+          text: "try again maybe?",
+          icon: "error",
+        })
+      );
+      
     }
   };
 
   logout = () => {
-    localStorage.clear();
-    this.setState({
-      email: "",
-      password: "",
-      token: null,
+    Swal.fire({
+      title: "Logout?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes !",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        localStorage.clear();
+        this.setState({
+          email: "",
+          password: "",
+          token: null,
+        });
+      }
     });
   };
 
@@ -212,6 +267,7 @@ class Navigation extends Component {
           text: "we are can't search empty string",
           icon: "error",
         });
+    this.setState({ searchInput: "" });
   };
 
   launchModalLogin = () => {
@@ -230,9 +286,19 @@ class Navigation extends Component {
     this.setState({ showSign: false });
   };
 
+  handleFileUpload = (event) => {
+    const file = event.currentTarget.files[0];
+    this.setState({
+      image: {
+        file: event.currentTarget.files[0],
+        url: URL.createObjectURL(file),
+      },
+    });
+  };
+
   render() {
-    const { token, loading } = this.state;
-    const usernameLog = this.state.dataLoggedIn.username;
+    const { token, loading, searchInput,userLogin } = this.state;
+    // const usernameLog = this.state.dataLoggedIn.username;
 
     return (
       <>
@@ -244,13 +310,13 @@ class Navigation extends Component {
               </Link>
             </Navbar.Brand>
 
-            {/* <InputGroup className="w-50"> */}
             <Form inline onSubmit={this.search}>
               <FormControl
                 placeholder="Search Movie..."
                 aria-label="Search Movie..."
                 aria-describedby="basic-addon2"
                 onChange={this.searchApa}
+                value={searchInput}
               />
               <Button variant="outline-secondary" type="submit">
                 Search
@@ -262,7 +328,7 @@ class Navigation extends Component {
               <Nav className="ml-auto">
                 {token ? (
                   <>
-                    <Button>Hello {usernameLog}</Button>
+                    <Link to="/user">Hello {userLogin}</Link>
                     <Button onClick={this.logout} variant="light">
                       logout
                     </Button>
@@ -299,6 +365,7 @@ class Navigation extends Component {
             <Formik
               validationSchema={schema}
               initialValues={{
+                image: "",
                 email: "",
                 password: "",
                 fullname: "",
@@ -312,16 +379,47 @@ class Navigation extends Component {
                 }, 400);
               }}
             >
-              {({
-                touched,
-                errors,
-                isSubmitting,
-                handleChange,
-                handleSubmit,
-              }) => (
+              {({ touched, errors, handleChange, handleSubmit }) => (
                 <Form
                   onSubmit={handleSubmit} // ini handle submitnya formik
                 >
+                  {/* <Form.Group controlId="Image">
+                    <Form.File>
+                      <div className="preview-pp">
+                        {this.state.image.url && (
+                          <Image
+                            roundedCircle
+                            fluid
+                            // width="350"
+                            // fill
+                            src={this.state.image.url}
+                            alt={this.state.image.file.name}
+                          />
+                        )}
+                      </div>
+                      <div className="preview-name">
+                        {this.state.image.file.name && (
+                          <span>{this.state.image.file.name}</span>
+                        )}
+                      </div>
+                      <Field
+                        type="file"
+                        placeholder="Enter your Fullname"
+                        name="image"
+                        // value={this.state.fullname}
+                        className={`form-control ${
+                          touched.image && errors.image ? "is-invalid" : "" // ini biar si field ada class form control dan invalid apa ga
+                        }`}
+                        onChange={(e) => this.handleFileUpload(e)}
+                        // onChange={handleChange}
+                      />
+                      <ErrorMessage
+                        component="div"
+                        name="image"
+                        className="invalid-feedback"
+                      />
+                    </Form.File>
+                  </Form.Group> */}
                   <Form.Group controlId="fullname">
                     <Form.Label>Full Name</Form.Label>
                     <Field
@@ -412,21 +510,25 @@ class Navigation extends Component {
                       className="invalid-feedback"
                     />
                   </Form.Group>
-                  <div className="align-items-center">
-                    <Button
-                      variant="primary"
-                      type="submit"
-                      disabled={isSubmitting}
-                    >
-                      {isSubmitting ? "please wait..." : "Submit"}
-                    </Button>
-                    <p>
-                      Already have an Account?{" "}
-                      <a href="#" onClick={this.openLogin}>
-                        Login
-                      </a>
-                    </p>
-                  </div>
+                  <Row>
+                    <Col md={10}>
+                      <p>
+                        Already have an Account?{" "}
+                        <a href="#" onClick={this.openLogin}>
+                          Login
+                        </a>
+                      </p>
+                    </Col>
+                    <Col>
+                      <Button
+                        variant="primary"
+                        type="submit"
+                        disabled={loading}
+                      >
+                        {loading ? "Loading... " : "Submit"}
+                      </Button>
+                    </Col>
+                  </Row>
                 </Form>
               )}
             </Formik>
@@ -502,17 +604,16 @@ class Navigation extends Component {
                       className="invalid-feedback"
                     />
                   </Form.Group>
-                  <div className="align-items-center">
-                    <Button variant="primary" type="submit" disabled={loading}>
-                      {loading ? "please wait..." : "Submit"}
-                    </Button>
-                    <p>
-                      Dont have an account yet?{" "}
-                      <a href="#" onClick={this.openSign}>
-                        Sign up
-                      </a>
-                    </p>
-                  </div>
+
+                  <p>
+                    Dont have an account yet?{" "}
+                    <a href="#" onClick={this.openSign}>
+                      Sign up
+                    </a>
+                  </p>
+                  <Button variant="primary" type="submit" disabled={loading}>
+                    {loading ? "please wait..." : "Submit"}
+                  </Button>
                 </Form>
               )}
             </Formik>
